Add TimeFrame_V4_0_full.ino
This a complete re-write of the sketch with the following changes:

   - More comments to help reader understand how it works
   - More use of named constants for readability
   - More use of functions for readability
   - Better constant and variable names
   - Unnecessary code removed
   - On-board LED heartbeat is now always on, but code rewritten to
     not interfere with timing
   - Implementation of Paul Hutchison's improvements:
       - Glowing LED for pushbutton (OPTIONAL)
       - Frame starts in standby mode
       - Removed LED only mode
       - Additional analog input to control electromagnet duty
         (OPTIONAL)
       - Adjustments to ranges
   - NEW FEATURE: frequency control

Note that this sketch is compatible with with both the original hardware
design and with either or both of the hardware modifications added by
Paul Hutchison (pushbutton LED and electromagnet strength knob.)
However, the constants HW_HAS_PUSHBUTTON_LED and
HW_HAS_MAG_STRENGTH_KNOB must be set appropriately below.

The frequency control feature uses the same knob (potentiometer) as the
LED brightness. The ability to vary the frequency allows the user to
find the resonant frequency of the object(s). Frequency control mode is
entered by "double clicking" the pushbutton when it is in the standby
mode. The LED brightness value is captured on entry to the frequency
control mode, and the knob is temporarily re-purposed to control the
frequency. The speed control knob and the electromagnet strength knob
work as normal while in frequency control mode. Frequency control mode
is exited and normal slow motion mode is entered when the pushbutton is
pressed. At this transition, the frequency value is captured and
continues to be used while the frame is powered on. Additionally, the
value is written to EEPROM. This value is read the next time the frame
is powered on and is used until the user enters frequency control mode
again to change it.

In frequency control mode, the pushbutton LED blinks rapidly. On
hardware that does not implement the pushbutton LED, the only way to
tell that it is in frequency control mode is to go ahead and turn the
knob that normally controls the brightness and see if it instead is
controlling the frequency.